+++ src/main/java/kr/co/itresumeregistersite/domain/board/dto/EditPostDto.java
@@ -1,0 +1,24 @@
+package kr.co.itresumeregistersite.domain.board.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+import javax.validation.constraints.NotBlank;
+import java.time.LocalDateTime;
+
+@Getter
+@Builder
+@AllArgsConstructor
+public class EditPostDto {
+
+    private Long boardId;
+
+    @NotBlank(message = "제목을 작성해주세요")
+    private String title;
+
+    @NotBlank(message = "내용을 작성해주세요")
+    private String content;
+
+    private LocalDateTime modifiedDate;
+}
