+++ src/main/java/kr/co/itresumeregistersite/domain/user/dto/UpdatePasswordDto.java
@@ -1,0 +1,19 @@
+package kr.co.itresumeregistersite.domain.user.dto;
+
+import lombok.Getter;
+
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.Size;
+
+@Getter
+public class UpdatePasswordDto { // 회원 비밀번호 수정 Dto
+
+    private String identity;
+
+    @NotBlank(message = "기존 비밀번호를 입력해주세요")
+    private String password;
+
+    @NotBlank(message = "변경할 비밀번호를 입력해주세요")
+    @Size(min = 4, max = 15, message = "PW는 4~15자로 입력해주세요")
+    private String changePassword;
+}
