--- src/main/java/kr/co/itresumeregistersite/repository/board/BoardRepository.java
@@ -1,14 +1,0 @@
-package kr.co.itresumeregistersite.repository.board;
-
-import kr.co.itresumeregistersite.domain.entity.board.Board;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface BoardRepository extends JpaRepository<Board, Long> {
-
-    List<Board> findByTitle(String title, Pageable pageable);
-}
